+++ packages/Theming/src/index.ts
@@ -1,13 +1,18 @@
 import { styled } from './styled'
 import { WHITE, BLACK, BLUE, INDIGO } from './colors'
 
+/** Style helpers */
 export * from './provider'
 export * from './colors'
 export * from './styled'
 export * from './styles'
+export * from './server'
+
+/** Color manipulation */
 export * from './rgba'
 export * from './shade'
-export * from './server'
+export * from './hsl'
+export * from './hex'
 
 export const lightTheme = {
   text: BLACK,
